+++ docs/.vitepress/theme/Layout.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+import DefaultTheme from 'vitepress/theme'
+import SiteFooterLinks from './components/SiteFooterLinks.vue'
+
+const { Layout } = DefaultTheme
+</script>
+
+<template>
+  <Layout>
+    <template #layout-bottom>
+      <SiteFooterLinks />
+    </template>
+  </Layout>
+</template>